ai-agent-template/requirements.txt
langchain
langchain-community
langgraph
llama-cpp-python
python-dotenv
chromadb
pypdf
